Use redux toolkit for settings store
