+++ src/store/SettingsStore/SettingsHooks.ts
@@ -1,24 +1,14 @@
 import { useMemo } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { AppState } from "..";
-import { setThemeAction } from "./SettingsActions";
+import { AppDispatch, AppState } from "..";
+import { setThemeAction } from "./SettingsSlice";
 import { themes } from "./Themes";
 
 const themeSelector = (state: AppState) => state.settingsStore.theme;
 
 export const useTheme = () => {
-  const dispatch = useDispatch();
+  const dispatch = useDispatch<AppDispatch>();
   const theme = useSelector(themeSelector);
-
-  /*
-  const toggleTheme = useCallback(() => {
-    if (theme === "light") {
-      dispatch(setThemeAction("dark"));
-    } else {
-      dispatch(setThemeAction("light"));
-    }
-  }, [dispatch, theme]);
-  */
 
   const colors = useMemo(() => {
     return themes[theme];
